requirements.txt
ultralytics>=8.0.0
gradio>=4.0.0
pillow>=10.0.0
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0